Added internal GLColor
Color now uses integer r, g, b and a values and is used in interface.
Closed #371

+++ src/gamebase/include/gamebase/drawobj/Atlas.h
@@ -24,7 +24,7 @@ public:
 	void setSize(const Vec2& size);
 
     Color color() const;
-    void setColor(float r, float g, float b, float a = 1);
+    void setColor(int r, int g, int b, int a = 1);
     void setColor(const Color& color);
 
     bool isVisible() const;
